### 2025-07-27 #8646 @editorialbot
[PRE REVIEW]: Vizagrams.jl: A Julia-based visualization framework that integrates diagramming and data visualization
**Submitting author:** @davibarreira (<a href="http://orcid.org/0000-0002-8005-4118">Davi Barreira</a>)
**Repository:** https://github.com/davibarreira/Vizagrams.jl
**Branch with paper.md** (empty if default branch): master
**Version:** v0.2.15
