+++ joss/paper.bib
@@ -7,6 +7,7 @@
     pages = {120--132},
     year = {2018},
     publisher = {Elsevier},
+    doi = {10.1016/j.jvlc.2017.10.001},
 }
 
 @article{d3js,
@@ -18,6 +19,7 @@
     pages = {2301--2309},
     year = {2011},
     publisher = {IEEE},
+    doi={10.1109/TVCG.2011.185}
 }
 
 @article{Hunter:2007,
@@ -69,6 +71,7 @@
     pages = {375--414},
     year = {2012},
     publisher = {Springer},
+    doi={10.1007/978-3-642-21551-3_13},
 }
 
 @article{wickham2007ggplot,
@@ -87,6 +90,7 @@
     pages = {180--185},
     year = {2011},
     publisher = {Wiley Online Library},
+    doi ={10.1002/wics.147},
 }
 @article{satyanarayan2016vega,
     title = {Vega-lite: A grammar of interactive graphics},
@@ -98,6 +102,7 @@
     pages = {341--350},
     year = {2016},
     publisher = {IEEE},
+    doi = {10.1109/TVCG.2016.2599030},
 }
 @misc{satyanarayan2014vega,
     author = {Satyanarayan, Arvind and Moritz, Dominik and Wongsuphasawat, Kanit
@@ -138,14 +143,15 @@
                  Functional Art, Music, Modelling and Design},
     pages = {4--5},
     year = {2015},
+    doi = {10.1145/2808083.2808085},
 }
-
-@article{pollock2023bluefish,
-    title = {Bluefish: A Relational Grammar of Graphics},
-    author = {Pollock, Josh and Mei, Catherine and Huang, Grace and Jackson,
-              Daniel and Satyanarayan, Arvind},
-    journal = {arXiv preprint arXiv:2307.00146},
-    year = {2023},
+@inproceedings{pollock2024bluefish,
+  title={Bluefish: Composing diagrams with declarative relations},
+  author={Pollock, Josh and Mei, Catherine and Huang, Grace and Evans, Elliot and Jackson, Daniel and Satyanarayan, Arvind},
+  booktitle={Proceedings of the 37th Annual ACM Symposium on User Interface Software and Technology},
+  pages={1--21},
+  year={2024}
+  doi = {10.1145/3654777.3676465},
 }
 @misc{granstrom2022diagrammar,
     title = {Diagrammar: Simply Make Interactive Diagrams},
@@ -167,6 +173,7 @@
     pages = {52--65},
     year = {2002},
     publisher = {IEEE},
+    doi = {10.1109/2945.981851},
 }
 @inproceedings{pu2020probabilistic,
     title = {A probabilistic grammar of graphics},
@@ -175,6 +182,7 @@
                  Computing Systems},
     pages = {1--13},
     year = {2020},
+    doi = {10.1145/3313831.3376466}, 
 }
 @article{lyi2021gosling,
     title = {Gosling: A grammar-based toolkit for scalable and interactive
@@ -187,6 +195,7 @@
     pages = {140--150},
     year = {2021},
     publisher = {IEEE},
+    doi = {10.1109/tvcg.2021.3114876},
 }
 @misc{observable_plot,
     title = {Observable Plot},
@@ -201,14 +210,16 @@
     pages = {171--175},
     year = {2021},
     organization = {IEEE},
+    doi = {10.1109/vis49827.2021.9623315},
 }
 @inproceedings{improvise,
-    title = {Building highly-coordinated visualizations in improvise},
+    title = {Building highly-coordinated visualizations in Improvise},
     author = {Weaver, Chris},
     booktitle = {IEEE Symposium on Information Visualization},
     pages = {159--166},
     year = {2004},
     organization = {IEEE},
+    doi = {10.1109/infvis.2004.12},
[…]
