### 2025-07-30 fc5188f
:book: Added other related works mentioned by the JOSS review.
