+++ joss/paper.bib
@@ -150,8 +150,8 @@
   author={Pollock, Josh and Mei, Catherine and Huang, Grace and Evans, Elliot and Jackson, Daniel and Satyanarayan, Arvind},
   booktitle={Proceedings of the 37th Annual ACM Symposium on User Interface Software and Technology},
   pages={1--21},
-  year={2024}
-  doi = {10.1145/3654777.3676465},
+  year={2024},
+  doi={10.1145/3654777.3676465},
 }
 @misc{granstrom2022diagrammar,
     title = {Diagrammar: Simply Make Interactive Diagrams},
@@ -273,6 +273,7 @@
     volume = {107},
     pages = {1--32},
     year = {2023},
+    doi={10.18637/jss.v107.i04}
 }
 @article{wickham2011split,
     title = {The split-apply-combine strategy for data analysis},
@@ -294,6 +295,7 @@
     pages = {144--1},
     year = {2020},
     publisher = {ACM New York, NY, USA},
+    doi = {10.1145/3386569.3392375},
 }
 @article{yorgey2012monoids,
     title = {Monoids: theme and variations (functional pearl)},
@@ -304,6 +306,7 @@
     pages = {105--116},
     year = {2012},
     publisher = {ACM New York, NY, USA},
+    doi={10.1145/2430532.2364520}
 }
 @software{plotsjl,
     author = {Tom Breloff},
@@ -327,4 +330,28 @@
     year = {2025},
     url = {https://hdl.handle.net/10438/36341},
     note = {PhD thesis},
+}
+@article{gramm,
+    doi = {10.21105/joss.00568},
+    url = {https://doi.org/10.21105/joss.00568},
+    year = {2018},
+    publisher = {The Open Journal},
+    volume = {3},
+    number = {23},
+    pages = {568},
+    author = {Morel, Pierre},
+    title = {Gramm: grammar of graphics plotting in Matlab},
+    journal = {Journal of Open Source Software},
+}
+@article{ggalluvial,
+    doi = {10.21105/joss.02017},
+    url = {https://doi.org/10.21105/joss.02017},
+    year = {2020},
+    publisher = {The Open Journal},
+    volume = {5},
+    number = {49},
+    pages = {2017},
+    author = {Brunson, Jason Cory},
+    title = {ggalluvial: Layered Grammar for Alluvial Plots},
+    journal = {Journal of Open Source Software},
 }
