+++ .github/workflows/CI.yml
@@ -33,7 +33,7 @@ jobs:
           - x64
           # - x86
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v5
       - uses: julia-actions/setup-julia@v2
         with:
           version: ${{ matrix.version }}
