### 2025-09-14 #8646 @jbytecode
👋👋👋 Dear @simondanisch and @gdalle 👋👋👋

**[[PRE REVIEW]: Vizagrams.jl: A Julia-based visualization framework that integrates diagramming and data visualization](https://github.com/openjournals/joss-reviews/issues/8646#top)**

You can find more information at the top of this Github issue ([LINK](https://github.com/openjournals/joss-reviews/issues/8646)).

Thank you in advance!
