### 2025-09-14 #8646 @SimonDanisch
Hi, I don't think I'd be the best reviewer for this since my area of expertise is rather in low level programming for graphics, and not as much  about making nice DSLs for it ;)
On top I'm also pretty swamped this year :(
Best of luck!
