### 2025-09-20 #8646 @abhishektiwari
@jbytecode happy to help with this submission. Looks quite straightforward.
