### 2025-09-21 davibarreira/Vizagrams.jl#90 @abhishektiwari
[JOSS Review] Add contribution guidelines
> Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3.

Please add contribution guideline and to mention/link to them in your README. You could for instance create a CONTRIBUTING.md file (see here for some examples: https://contributing.md/example/).

For [JOSS Review](https://github.com/openjournals/joss-reviews/issues/9037)
