### 2025-09-21 #9037 @abhishektiwari
@davibarreira - can you clarify if [authorship](https://joss.readthedocs.io/en/latest/submitting.html#authorship) of paper is appropriate and complete? I don't see code contribution from other 2 authors on paper.
