### 2025-09-22 #9037 @davibarreira
@abhishektiwari , the software was part of my phd thesis, and the two authors present besides me were my advisors. Although they did not directly submit any code, both of them were essential for the creation of the package.

About the two who appear as contributors in the GitHub page for the software: one of them helped with some grammatical/formatting errors in the docs, and the other added a single line of code to catch some NaNs values.
I thought their contributions, although very welcomed, were not sufficient for including them as authors in the paper.
