### 2025-09-22 037ffbc
:truck: Adding contributions guideline to the README.md
