### 2025-10-03 eca298b
:books: paper.bib corrections for JOSS.
