+++ joss/paper.bib
@@ -19,7 +19,7 @@
     pages = {2301--2309},
     year = {2011},
     publisher = {IEEE},
-    doi={10.1109/TVCG.2011.185}
+    doi = {10.1109/TVCG.2011.185},
 }
 
 @article{Hunter:2007,
@@ -71,14 +71,13 @@
     pages = {375--414},
     year = {2012},
     publisher = {Springer},
-    doi={10.1007/978-3-642-21551-3_13},
+    doi = {10.1007/978-3-642-21551-3_13},
 }
 
 @article{wickham2007ggplot,
     title = {The ggplot package},
     author = {Wickham, Hadley and Wickham, Maintainer Hadley},
-    journal = {URL: https://cran. r-project. org/web/packages/ggplot2/index.
-               html},
+    journal = {URL: https://cran.r-project.org/web/packages/ggplot2/index.html},
     year = {2007},
 }
 @article{wickham2011ggplot2,
@@ -90,7 +89,7 @@
     pages = {180--185},
     year = {2011},
     publisher = {Wiley Online Library},
-    doi ={10.1002/wics.147},
+    doi = {10.1002/wics.147},
 }
 @article{satyanarayan2016vega,
     title = {Vega-lite: A grammar of interactive graphics},
@@ -108,7 +107,7 @@
     author = {Satyanarayan, Arvind and Moritz, Dominik and Wongsuphasawat, Kanit
               and Heer, Jeffrey},
     title = {Vega: A Visualization Grammar},
-    howpublished = {\url{http://trifacta.github.io/vega}},
+    howpublished = {http://trifacta.github.io/vega},
     month = {April},
     year = {2014},
 }
@@ -146,12 +145,14 @@
     doi = {10.1145/2808083.2808085},
 }
 @inproceedings{pollock2024bluefish,
-  title={Bluefish: Composing diagrams with declarative relations},
-  author={Pollock, Josh and Mei, Catherine and Huang, Grace and Evans, Elliot and Jackson, Daniel and Satyanarayan, Arvind},
-  booktitle={Proceedings of the 37th Annual ACM Symposium on User Interface Software and Technology},
-  pages={1--21},
-  year={2024},
-  doi={10.1145/3654777.3676465},
+    title = {Bluefish: Composing diagrams with declarative relations},
+    author = {Pollock, Josh and Mei, Catherine and Huang, Grace and Evans,
+              Elliot and Jackson, Daniel and Satyanarayan, Arvind},
+    booktitle = {Proceedings of the 37th Annual ACM Symposium on User Interface
+                 Software and Technology},
+    pages = {1--21},
+    year = {2024},
+    doi = {10.1145/3654777.3676465},
 }
 @misc{granstrom2022diagrammar,
     title = {Diagrammar: Simply Make Interactive Diagrams},
@@ -182,12 +183,12 @@
                  Computing Systems},
     pages = {1--13},
     year = {2020},
-    doi = {10.1145/3313831.3376466}, 
+    doi = {10.1145/3313831.3376466},
 }
 @article{lyi2021gosling,
     title = {Gosling: A grammar-based toolkit for scalable and interactive
              genomics data visualization},
-    author = {LYi, Sehi and Wang, Qianwen and Lekschas, Fritz and Gehlenborg,
+    author = {LY'i, Sehi and Wang, Qianwen and Lekschas, Fritz and Gehlenborg,
               Nils},
     journal = {IEEE Transactions on Visualization and Computer Graphics},
     volume = {28},
@@ -243,7 +244,7 @@
                  Computing Systems},
     pages = {1--13},
     year = {2018},
-    doi = {10.1145/3173574.3173697}, 
[…]
