### 2025-10-03 9b282ae
:books: paper.bib corrections for JOSS.
