+++ joss/paper.bib
@@ -74,7 +74,7 @@
     doi = {10.1007/978-3-642-21551-3_13},
 }
 
-@article{wickham2007ggplot,
+@software{wickham2007ggplot,
     title = {The ggplot package},
     author = {Wickham, Hadley and Wickham, Maintainer Hadley},
     journal = {URL: https://cran.r-project.org/web/packages/ggplot2/index.html},
