### 2025-10-07 50ebb6a
:bug: Fixing the error message for Hist mark.
