### 2025-10-12 1220923
add citation to Julia
