+++ joss/paper.bib
@@ -356,3 +356,19 @@
     title = {ggalluvial: Layered Grammar for Alluvial Plots},
     journal = {Journal of Open Source Software},
 }
+
+
+@article{julia,
+  doi       = {10.1137/141000671},
+  url       = {https://doi.org/10.1137%2F141000671},
+  year      = 2017,
+  month     = {jan},
+  publisher = {Society for Industrial {\&} Applied Mathematics ({SIAM})},
+  volume    = {59},
+  number    = {1},
+  pages     = {65--98},
+  author    = {Jeff Bezanson and Alan Edelman and Stefan Karpinski and Viral B. Shah},
+  title     = {{J}ulia: A Fresh Approach to Numerical Computing},
+  journal   = {{SIAM} Review}
+}
+
