### 2025-10-12 e2cf990
fix capitalizations in bibtex
