+++ joss/paper.bib
@@ -60,8 +60,7 @@
     number = {65},
     pages = {3349},
     author = {Simon Danisch and Julius Krumbiegel},
-    title = {{Makie.jl}: Flexible high-performance data visualization for {Julia
-             }},
+    title = {{Makie.jl}: Flexible high-performance data visualization for {J}ulia},
     journal = {Journal of Open Source Software},
 }
 @incollection{wilkinson2012grammar,
@@ -129,10 +128,10 @@
     url = {https://doi.org/10.5281/zenodo.5559613},
 }
 @misc{composejl,
-    title = {Compose.jl: Declarative Vector Graphics for Julia},
+    title = {Compose.jl: Declarative Vector Graphics for {J}ulia},
     author = {Giovine, Samuel},
     year = {2022},
-    url = {\url{https://github.com/GiovineItalia/Compose.jl}},
+    url = {https://github.com/GiovineItalia/Compose.jl},
     note = {Accessed: March 11, 2024},
 }
 @inproceedings{yates2015diagrams,
@@ -268,7 +267,7 @@
     doi = {10.1145/3313831.3376297},
 }
 @article{bouchet2023dataframes,
-    title = {Dataframes. jl: Flexible and fast tabular data in Julia},
+    title = {{Dataframes.jl}: Flexible and fast tabular data in {J}ulia},
     author = {Bouchet-Valat, Milan and Kami{\'n}ski, Bogumi{\l}},
     journal = {Journal of Statistical Software},
     volume = {107},
@@ -341,7 +340,7 @@
     number = {23},
     pages = {568},
     author = {Morel, Pierre},
-    title = {Gramm: grammar of graphics plotting in Matlab},
+    title = {Gramm: grammar of graphics plotting in {M}atlab},
     journal = {Journal of Open Source Software},
 }
 @article{ggalluvial,
