+++ Project.toml
@@ -1,7 +1,7 @@
 name = "Vizagrams"
 uuid = "8c229dad-8b3a-4031-83d6-73545c88426d"
 authors = ["Davi Barreira <davisbarreira@gmail.com> and contributors"]
-version = "0.2.15"
+version = "0.2.16"
 
 [deps]
 Accessors = "7d9f7c33-5ae7-4f3b-8dc6-eff91059b697"
